refactor(webreel): extract shared compositeRecording helper
runVideo and the composite command each re-implemented the
zoom-filter/compose/thumbnail sequence, and the two copies had already
drifted (runner.ts had verbose per-event logging the command lacked).
Move normalizeAutoZoom and extractThumbnailIfConfigured into a new
lib/compositing.ts alongside a new compositeRecording() helper, and
have runVideo call it. Both names are re-exported from runner.ts for
compatibility with existing importers.
